refactor: organize attribute classes in folders per section
